requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pyserial==3.4
requests==2.21.0
urllib3==1.24.3
